fix: changes language for global/group/private tooltips and global icon My intuition for privacy was (from least to most private): Everyone -> some people -> just me however, the icons were: open eye -> group of people -> closed eye and the tooltips were: other users -> groups -> not visible
All that didn't make a whole lot of intuitive sense to me,
and made even less sense after looking at the code for it.

This changes it to be:
icons
globe -> some people -> closed eye
tooltip
public to district -> shared with groups -> visible only to me

Borrowing some of facebook's terminology with only me and public,
in addition to making the sentences different looking that a casual
(and possibly not wearing glasses) notices their difference.