refactor: adds pagination component pulls Josh's renderPagination function out of VJS and puts it into it's own component.
After woefully breaking storybook, I learned that pagination is not free
when it comes to perf, adding a few thousand pages of results will break
the page from the paginator alone. So the function has been changed
to generate an array of only nearby values reducting complexity from
O(N) to O(1) allowing values upto 15 or so digits, which will then cause
an overflow in the key of the javascript array.

TIL: `git mv -f OldFileNameCase newfilenamecase`