refactor: fixes wonky AppContainer For some reason the AppContainer component was larger than the shell, this had the effect of making scroll work a little strangely and occasionally elements would appear outside of the shell, have two scroll bars, or look odd.
First, we put the AppContainer in a styled div with restricted height,
now the app is contained within the shell, everything above the app is truly
above and everything below is truly below.

Because every module thusfar has assumed a wonky AppContainer, we must shift
the header elements down as to not be too close to the NavBar.

Also, certain elements are too close to the CallBar on the bottom of the page
to remedy that, we notice that most components are in AntD Rows with an
mb-3 class added too them except the bottom rows. Going through and adding
mb-3 classes to the bottom of every module seems silly, so we add the class
to the vjs card itself, nothing changes about the rows, and the cards on the
bottom stay far enough from the CallBar to not make anyone cry.

We also need to change the height of the Content portion of the tagbuilder,
because we changed the height of AppContainer, and it needs to match.

Last but not least we remove all the mb-60 classes because we should never
have to reference the bottom of the page anymore, the modules should be
contained in containers that end at the CallBar rather than beneath it.