refactor: improve SMS read status settings
Fix a bug where the message only appeared to be unread if the last
message in the conversation was inbound and unread. This could cause a
problem if someone is using the mobile app _and_ ss-ui, leading to a
message counter that wouldn't seem to go away without any visual
indication of unread messages.

Also change it so that whenever a conversation is clicked, all logs in
that convo are marked as read. This is what the mobile app does right
now.