fix: disappearing selection for 'is any of'
Problem:
In order for QueryBuilder to work, it expects an `OrderedMap` from
Immutable JS(https://facebook.github.io/immutable-js/docs/#/).
To convert the treeQuery,  we pass it to a utility function called `fromJSToOrderedMap`. This function goes through all the values of the treeQuery and converts any object to an `OrderedMap` and any Array to a `List`.

Soln:
The problem is that for multiselect fields like `select_any_in` or
`select_not_any_in` they expect an Array instead of a `List`. This is
why whenever those fields were selected, the value would be empty.

solves #565