fix: save selected groups
Issue:
`this.groups` object was expected to have an array of group objects with
`id`. To retrieve `groupIds` you would transform the group
using a `map` and only get the ids: `this.groups.map(g => g.id). The
issue was when a user selected multiple groups from the ui, they would
be saved in the `groups` object as an array of `groupId` strings instead
of an array of group objects.

Solution:
Make sure groups array only contains group objects.

solves #662