fix: edit group scope
Problem
u were editing the scope of a Tag to either [Just Me,
Everyone], the `groups` array would be converted into an empty string.
Strings do not have a `map` function as a result, this caused the app to
crash.

Solution:
Make sure `groups` is always an array either empty or with groups.

solves #709