:bug:: remove deleted groups from list
Remove recently created groups from the group list.

This bug was caused by the fact that the UUID generated on the client
side prior to actually saving on the server was the key in the `groups`
map but wasn't what the delete function is looking to actually delete.
This fix is in line with how notes handles the same situation.

Fixes #737